Walmart Weekly Sales Analysis
In this Jupyter Notebook, we will perform an analysis of Walmart sales data for 45 stores between the period of 2010-02-05 to 2012-10-26. The dataset includes various features such as store information, department details, economic indicators like CPI and unemployment rate, weather information, promotional markdowns, fuel prices, and more. The main objective is to predict the weekly sales based on these features.